// src/app/challenge/shared/model/tipo-uf.enum.ts
export enum TipoUf {
    pe = 'PE',
    sp = 'SP',
    pr = 'PR',
    pb = 'PB',
    rj = 'RJ',
    ce = 'CE',
    ac = 'AC',
    al = 'AL',
    ap = 'AP',
    am = 'AM',
    ba = 'BA',
    df = 'DF',
    es = 'ES',
    go = 'GO',
    ma = 'MA',
    mt = 'MT',
    ms = 'MS',
    mg = 'MG',
    pa = 'PA',
    pi = 'PI',
    rn = 'RN',
    rs = 'RS',
    ro = 'RO',
    rr = 'RR',
    sc = 'SC',
    se = 'SE',
    to = 'TO',
}

export const listaUfs =['AC', 'AL', 'AP', 'AM', 'BA',
    'CE', 'DF', 'ES', 'GO', 'MA', 'MT', 'MS', 'MG', 'PA', 'PB', 'PR',
    'PE', 'PI', 'RJ', 'RN', 'RS', 'RO', 'RR', 'SC', 'SP', 'SE', 'TO'];